remove models and rename file
