+++ src/api/model/mod.rs
@@ -3,6 +3,6 @@ pub mod ready;
 pub mod play;
 pub mod gateway;
 pub mod track;
-pub mod voice;
+pub mod nz;
 pub mod info;
 pub mod player;
